// src/types/MessageResponse.ts
export interface MessageResponse {
    FLAGS?: number
    FROMSYMBOL?: string
    LASTTRADEID?: string
    LASTUPDATE?: number
    LASTVOLUME?: number
    LASTVOLUMETO?: number
    MARKET?: string
    PRICE?: number
    TOSYMBOL?: string
    TYPE?: string
    VOLUME24HOUR?: number
    VOLUME24HOURTO?: number
    VOLUMEDAY?: number
    VOLUMEDAYTO?: number
    VOLUMEHOUR?: number
    VOLUMEHOURTO?: number
}